Update to deploy

+++ src/documents/services/documents.service.ts
@@ -238,12 +238,14 @@ export class DocumentsService {
         )
       }
 
-      if (document) {
+      if (document && document.id) {
         await this.documentsRepository.delete(document.id)
       }
       if (numeration) {
         await this.numerationDocumentService.remove(numeration.id)
       }
+
+      console.error(error)
 
       throw new Error(error.message)
     }
